move /project pages outside of _appshell layout
the project form does not follow the _appshell layout so it should be
separated
